Reorganize scheduling plugins
